Establish numeric types and arithmetic standard (#1120)
* Establish numeric types and arithmetic standard

* Keep numeric implementation details out of issue templates

* Move SAT allocation into rule helpers

* Tighten structural semantic review

* Revert structural review wording

+++ problemreductions-cli/src/commands/create/schema_support.rs
@@ -456,7 +456,7 @@ pub(super) fn resolve_schema_field_type(
 
 pub(super) fn weight_sum_type(weight_type: &str) -> &'static str {
     match weight_type {
-        "One" | "i32" => "i32",
+        "One" | "i32" => "i64",
         "f64" => "f64",
         _ => "i32",
     }
